requirements.txt
gcloud
google-auth
requests
earthengine-api